+++ docs/changelog.md
@@ -4,6 +4,13 @@
 
 # Changelog
 
+## 0.71.0 (July 2025)
+
+- Update minimum pytket version to 2.8.0.
+- Update minimum qiskit version to 2.1.1.
+- Document limitations of conversion of symbolic circuits between pytket and
+  qiskit.
+
 ## 0.70.0 (July 2025)
 
 - Fix overwriting of user-supplied `service` parameter to `IBMQBackend`.